Remove ORPC page_cache.
This replaces the page cache with the "baseline" version. This
effectively rolls back the page cache to the upstream version.
