+++ kernel/src/fs/fs_impls/ext2/block_group.rs
@@ -1,11 +1,6 @@
 // SPDX-License-Identifier: MPL-2.0
 
-#[cfg(not(baseline_asterinas))]
-use ostd::orpc::{
-    oqueue::{OQueue as _, OQueueRef},
-    orpc_impl,
-};
-use ostd::{const_assert, mm::io::util::HasVmReaderWriter, new_server, orpc::orpc_server};
+use ostd::{const_assert, mm::io::util::HasVmReaderWriter};
 
 use super::{
     block_ptr::Ext2Bid,
@@ -14,9 +9,10 @@ use super::{
     prelude::*,
     super_block::SuperBlock,
 };
-use crate::fs::utils::IdBitmap;
-#[cfg(not(baseline_asterinas))]
-use crate::fs::vfs::server_traits::{self, PageIOObservable, PageStore};
+use crate::fs::{
+    utils::IdBitmap,
+    vfs::page_cache::{CachePage, PageCacheBackend},
+};
 
 /// Blocks are clustered into block groups in order to reduce fragmentation and minimise
 /// the amount of head seeking when reading a large amount of consecutive data.
@@ -26,7 +22,6 @@ pub(super) struct BlockGroup {
     raw_inodes_cache: PageCache,
 }
 
-#[orpc_server(server_traits::PageStore, server_traits::PageIOObservable)]
 struct BlockGroupImpl {
     inode_table_bid: Ext2Bid,
     raw_inodes_size: usize,
@@ -87,7 +82,7 @@ impl BlockGroup {
                 }
             };
 
-            new_server!(|_| BlockGroupImpl {
+            Arc::new(BlockGroupImpl {
                 inode_table_bid: metadata.descriptor.inode_table_bid,
                 raw_inodes_size,
                 inner: RwMutex::new(Inner {
@@ -331,75 +326,6 @@ impl Debug for BlockGroup {
     }
 }
 
-#[cfg(not(baseline_asterinas))]
-#[orpc_impl]
-impl PageIOObservable for BlockGroupImpl {
-    fn page_reads_oqueue(&self) -> OQueueRef<usize>;
-    fn page_reads_reply_oqueue(&self) -> OQueueRef<usize>;
-    fn page_writes_oqueue(&self) -> OQueueRef<usize>;
-    fn page_writes_reply_oqueue(&self) -> OQueueRef<usize>;
-}
-
-#[cfg(not(baseline_asterinas))]
-#[orpc_impl]
-impl PageStore for BlockGroupImpl {
-    fn read_page_async(&self, req: server_traits::AsyncReadRequest) -> Result<()> {
-        let bid = self.inode_table_bid + req.handle.idx as Ext2Bid;
-        // TODO: Should we allocate the bio segment from the pool on reads?
-        // This may require an additional copy to the requested frame in the completion callback.
-        let bio_segment = BioSegment::new_from_segment(
-            Segment::from(req.handle.frame.clone()).into(),
-            BioDirection::FromDevice,
-        );
-
-        let reply_producer = self.page_reads_reply_oqueue().attach_ref_producer()?;
-        self.page_reads_oqueue().produce_ref(&req.handle.idx)?;
-
-        self.fs
-            .upgrade()
-            .unwrap()
-            .read_blocks_async_with_closure(bid, bio_segment, move |_| {
-                // TODO(arthurp, #120): This can crash if produce blocks.
-                reply_producer.produce_ref(&req.handle.idx);
-                req.reply_handle.produce(req.handle);
-            })?;
-
-        Ok(())
-    }
-
-    fn write_page_async(&self, req: server_traits::AsyncWriteRequest) -> Result<()> {
-        let bid = self.inode_table_bid + req.handle.idx as Ext2Bid;
-        let bio_segment = BioSegment::alloc(1, BioDirection::ToDevice);
-        // This requires an additional copy to the pooled bio segment.
-        bio_segment
-            .writer()
-            .unwrap()
-            .write_fallible(&mut req.handle.frame.reader().to_fallible())?;
-
-        let reply_producer = self.page_reads_reply_oqueue().attach_ref_producer()?;
-        self.page_writes_oqueue().produce_ref(&req.handle.idx)?;
-
-        self.fs.upgrade().unwrap().write_blocks_async_with_closure(
-            bid,
-            bio_segment,
-            move |_| {
-                // TODO(arthurp, #120): This can crash if produce blocks.
-                reply_producer.produce_ref(&req.handle.idx);
-                if let Some(reply_handle) = req.reply_handle {
-                    reply_handle.produce(req.handle);
-                }
-            },
-        )?;
-
-        Ok(())
-    }
-
-    fn npages(&self) -> Result<usize> {
-        Ok(self.raw_inodes_size.div_ceil(BLOCK_SIZE))
-    }
-}
-
-#[cfg(baseline_asterinas)]
 impl PageCacheBackend for BlockGroupImpl {
     fn read_page_async(&self, idx: usize, frame: &CachePage) -> Result<BioWaiter> {
         let bid = self.inode_table_bid + idx as Ext2Bid;
