+++ kernel/src/fs/fs_impls/ext2/inode.rs
@@ -8,14 +8,8 @@ use core::sync::atomic::{AtomicUsize, Ordering};
 
 use device_id::DeviceId;
 use inherit_methods_macro::inherit_methods;
-#[cfg(not(baseline_asterinas))]
-use ostd::orpc::{
-    oqueue::{OQueue as _, OQueueRef},
-    orpc_impl,
-};
 use ostd::{
-    const_assert, mm::io::util::HasVmReaderWriter, new_server, orpc::orpc_server,
-    util::callback_counter::CallbackCounter,
+    const_assert, mm::io::util::HasVmReaderWriter, util::callback_counter::CallbackCounter,
 };
 
 use super::{
@@ -27,16 +21,13 @@ use super::{
     utils::now,
     xattr::Xattr,
 };
-#[cfg(baseline_asterinas)]
-use crate::fs::vfs::page_cache::PageCacheBackend;
-#[cfg(not(baseline_asterinas))]
-use crate::fs::vfs::server_traits::{self, PageIOObservable as _};
 use crate::{
     fs::{
         file::{InodeMode, Permission},
         pipe::Pipe,
         vfs::{
             inode::{Extension, FallocMode, Inode as _, Metadata},
+            page_cache::PageCacheBackend,
             path::{is_dot, is_dot_or_dotdot, is_dotdot},
             xattr::{XattrName, XattrNamespace, XattrSetFlags},
         },
@@ -992,18 +983,11 @@ impl InodeInner {
         let num_page_bytes = desc.num_page_bytes();
         let inode_impl = InodeImpl::new(desc, weak_self, fs);
         Self {
-            page_cache: {
-                let cache = PageCache::with_capacity(
-                    num_page_bytes,
-                    Arc::downgrade(&inode_impl.block_manager) as _,
-                )
-                .unwrap();
-                #[cfg(not(baseline_asterinas))]
-                {
-                    cache.start_prefetcher().unwrap();
-                }
-                cache
-            },
+            page_cache: PageCache::with_capacity(
+                num_page_bytes,
+                Arc::downgrade(&inode_impl.block_manager) as _,
+            )
+            .unwrap(),
             inode_impl,
         }
     }
@@ -1240,7 +1224,7 @@ struct InodeImpl {
 
 impl InodeImpl {
     pub fn new(desc: Dirty<InodeDesc>, weak_self: Weak<Inode>, fs: Weak<Ext2>) -> Self {
-        let block_manager = new_server!(|_| InodeBlockManager {
+        let block_manager = Arc::new(InodeBlockManager {
             nblocks: AtomicUsize::new(desc.blocks_count() as _),
             block_ptrs: RwMutex::new(desc.block_ptrs),
             indirect_blocks: RwMutex::new(IndirectBlockCache::new(fs.clone())),
@@ -1863,7 +1847,6 @@ impl InodeImpl {
 }
 
 /// Manages the inode blocks and block I/O operations.
-#[orpc_server(server_traits::PageStore, server_traits::PageIOObservable)]
 struct InodeBlockManager {
     nblocks: AtomicUsize,
     /// Maintains a second copy of block pointers for page cache use, distinct from
@@ -2049,48 +2032,6 @@ impl InodeBlockManager {
     }
 }
 
-#[cfg(not(baseline_asterinas))]
-#[orpc_impl]
-impl server_traits::PageIOObservable for InodeBlockManager {
-    fn page_reads_oqueue(&self) -> OQueueRef<usize>;
-    fn page_reads_reply_oqueue(&self) -> OQueueRef<usize>;
-    fn page_writes_oqueue(&self) -> OQueueRef<usize>;
-    fn page_writes_reply_oqueue(&self) -> OQueueRef<usize>;
-}
-
-#[cfg(not(baseline_asterinas))]
-#[orpc_impl]
-impl server_traits::PageStore for InodeBlockManager {
-    fn read_page_async(&self, req: server_traits::AsyncReadRequest) -> Result<()> {
-        let bid = req.handle.idx as Ext2Bid;
-        self.page_reads_oqueue().produce_ref(&req.handle.idx)?;
-        let reply_producer = self.page_reads_reply_oqueue().attach_ref_producer()?;
-        self.read_block_async_with_closure(bid, &req.handle.frame.clone(), move || {
-            // TODO(arthurp, #120): This can crash if produce blocks.
-            reply_producer.produce_ref(&req.handle.idx);
-            req.reply_handle.produce(req.handle);
-        })
-    }
-
-    fn write_page_async(&self, req: server_traits::AsyncWriteRequest) -> Result<()> {
-        let bid = req.handle.idx as Ext2Bid;
-        self.page_writes_oqueue().produce_ref(&req.handle.idx)?;
-        let reply_producer = self.page_writes_reply_oqueue().attach_ref_producer()?;
-        self.write_block_async_with_closure(bid, &req.handle.frame.clone(), move || {
-            // TODO(arthurp, #120): This can crash if produce blocks.
-            reply_producer.produce_ref(&req.handle.idx);
-            if let Some(reply_handle) = req.reply_handle {
-                reply_handle.produce(req.handle);
-            }
-        })
-    }
-
-    fn npages(&self) -> Result<usize> {
-        Ok(self.nblocks())
-    }
-}
-
-#[cfg(baseline_asterinas)]
 impl PageCacheBackend for InodeBlockManager {
     fn read_page_async(&self, idx: usize, frame: &CachePage) -> Result<BioWaiter> {
         let bid = idx as Ext2Bid;
