--- kernel/src/fs/vfs/server_traits.rs
@@ -1,215 +1,0 @@
-// SPDX-License-Identifier: MPL-2.0
-
-use alloc::sync::Arc;
-
-#[cfg(not(baseline_asterinas))]
-use ostd::orpc::oqueue::{
-    ConsumableOQueueRef, ElementDescriptor, GenericOQueueRef, OQueueRef, ValueProducer,
-    reply::ReplyQueue,
-};
-use ostd::orpc::orpc_trait;
-use serde::Serialize;
-
-#[cfg(not(baseline_asterinas))]
-use crate::event::EventContext;
-use crate::{fs::vfs::page_cache::CachePage, prelude::Result};
-
-/// A reference to a page in a [`PageStore`]. It contains the page index and the frame that holds
-/// the page data (if available).
-#[derive(Clone)]
-pub struct PageHandle {
-    /// The ID of the page (as an offset in the store).
-    pub idx: usize,
-    /// The page to read data from, or put data into.
-    pub frame: CachePage,
-}
-
-#[cfg(not(baseline_asterinas))]
-pub struct AsyncReadRequest {
-    pub handle: PageHandle,
-    /// A producer handle into an OQueue to send the reply to.
-    pub reply_handle: ValueProducer<PageHandle>,
-}
-
-#[cfg(not(baseline_asterinas))]
-pub struct AsyncWriteRequest {
-    pub handle: PageHandle,
-    /// A producer handle into an OQueue to send the reply to. If this is [`None`] no reply is sent.
-    pub reply_handle: Option<ValueProducer<PageHandle>>,
-}
-
-#[cfg(not(baseline_asterinas))]
-impl From<PageHandle> for AsyncWriteRequest {
-    fn from(handle: PageHandle) -> Self {
-        Self {
-            handle,
-            reply_handle: None,
-        }
-    }
-}
-
-#[cfg(not(baseline_asterinas))]
-fn new_oqueue<D: ElementDescriptor + 'static>() -> GenericOQueueRef<D> {
-    GenericOQueueRef::new_anonymous(8)
-}
-
-#[cfg(not(baseline_asterinas))]
-fn new_oqueue_with_len<D: ElementDescriptor + 'static>(len: usize) -> GenericOQueueRef<D> {
-    GenericOQueueRef::new_anonymous(len)
-}
-
-#[orpc_trait]
-pub trait PageIOObservable {
-    #[cfg(not(baseline_asterinas))]
-    fn page_reads_oqueue(&self) -> OQueueRef<usize> {
-        new_oqueue()
-    }
-
-    #[cfg(not(baseline_asterinas))]
-    fn page_reads_reply_oqueue(&self) -> OQueueRef<usize> {
-        // TODO: This must be longer than the largest number of IO that can be outstanding in the
-        // system. Otherwise a produce into this OQueue in the interrupt handler will block panicking
-        // the kernel.
-        new_oqueue_with_len(32)
-    }
-
-    #[cfg(not(baseline_asterinas))]
-    fn page_writes_oqueue(&self) -> OQueueRef<usize> {
-        new_oqueue()
-    }
-
-    #[cfg(not(baseline_asterinas))]
-    fn page_writes_reply_oqueue(&self) -> OQueueRef<usize> {
-        // TODO: as page_reads_reply_oqueue
-        new_oqueue_with_len(32)
-    }
-}
-
-/// A data store full of pages. This can be used to represent either a file or a block device.
-///
-/// NOTE: Read and write have both sync and async variants. Because of this, the implicit OQueues
-/// may not include all actual read operations. Use the explicit OQueues instead. The implementation
-/// should make sure to populate those OQueues.
-///
-/// TODO(arthurp, https://github.com/ldos-project/asterinas/issues/122): The trait takes an object
-/// oriented approach to match the existing file system implementation. It would be better to use a
-/// single server per filesystem and pass around richer page IDs than an index.
-#[orpc_trait]
-pub trait PageStore: PageIOObservable {
-    // TODO(arthurp, https://github.com/ldos-project/asterinas/issues/121): read_page_async and
-    // write_page_async should be OQueues, but doing so would make implementing them in the existing
-    // monolithic kernel code is tricky and not worth it at the moment.
-
-    // TODO(arthurp, https://github.com/ldos-project/asterinas/issues/121): Handle errors on async
-    // requests. They probably need to reply in the error case.
-
-    /// Reads a page asynchronously. The reply will be sent to [`AsyncReadRequest::reply_handle`].
-    #[cfg(not(baseline_asterinas))]
-    fn read_page_async(&self, handle: AsyncReadRequest) -> Result<()>;
-
-    /// Writes a page asynchronously. The reply will be sent to [`AsyncWriteRequest::reply_handle`]
-    /// if it is available.
-    #[cfg(not(baseline_asterinas))]
-    fn write_page_async(&self, handle: AsyncWriteRequest) -> Result<()>;
-
-    /// Reads a page synchronously.
-    #[cfg(not(baseline_asterinas))]
-    fn read_page(&self, handle: PageHandle) -> Result<()> {
-        let (reply_handle, consumer) = ReplyQueue::new_pair()?;
-        self.read_page_async(AsyncReadRequest {
-            handle,
-            reply_handle,
-        })?;
-        consumer.consume();
-        Ok(())
-    }
-
-    /// Writes a page synchronously.
-    #[cfg(not(baseline_asterinas))]
-    fn write_page(&self, handle: PageHandle) -> Result<()> {
-        let (reply_handle, consumer) = ReplyQueue::new_pair()?;
-        self.write_page_async(AsyncWriteRequest {
-            handle,
-            reply_handle: Some(reply_handle),
-        })?;
-        consumer.consume();
-        Ok(())
-    }
-
-    /// Returns the number of pages in this store.
-    fn npages(&self) -> Result<usize>;
-}
-
-/// The state of a page in the cache.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[repr(u8)]
-pub enum CacheState {
-    /// The page was in the cache.
-    Hit,
-    /// The page was not in the cache.
-    Miss,
-    /// The page was currently being read into the cache.
-    Pending,
-    Prefetch,
-    /// The page was evicted from the cache (LRU eviction).
-    #[expect(unused)]
-    Evict,
-}
-
-#[cfg(not(baseline_asterinas))]
-#[derive(Debug, Clone, Copy, Serialize)]
-pub struct PageCacheReadInfo {
-    pub idx: u64,
-    /// The state of the cached page when the request was made.
-    pub cache_state: CacheState,
-    pub cache_id: usize,
-    /// Total number of pages in the underlying store.
-    pub store_size: u64,
-    /// Number of pages currently held in the cache.
-    pub cache_pages: u64,
-    pub context: EventContext,
-}
-
-#[cfg(not(baseline_asterinas))]
-impl PageCacheReadInfo {
-    /// Creates a new PageCacheReadInfo with computed timestamp and tid.
-    pub fn new(
-        idx: u64,
-        cache_state: CacheState,
-        cache_id: usize,
-        store_size: u64,
-        cache_pages: u64,
-    ) -> Self {
-        PageCacheReadInfo {
-            idx,
-            cache_state,
-            cache_id,
-            store_size,
-            cache_pages,
-            context: EventContext::new(),
-        }
-    }
-}
-
-#[orpc_trait]
-pub trait PageCache {
-    #[cfg(not(baseline_asterinas))]
-    fn prefetch_oqueue(&self) -> ConsumableOQueueRef<usize> {
-        ConsumableOQueueRef::new_anonymous(8)
-    }
-
-    fn underlying_page_store(&self) -> Result<Arc<dyn PageStore>>;
-
-    #[cfg(not(baseline_asterinas))]
-    fn page_cache_read_info_oqueue(&self) -> OQueueRef<PageCacheReadInfo> {
-        new_oqueue_with_len(32)
-    }
-}
-
-#[orpc_trait]
-pub trait PageEvictionPolicy {
-    /// Get the page index to evict. The policy cannot decide, for whatever reason, this should
-    /// return an error. The caller will use some fallback. If the returned index does not refer to
-    /// a valid page in the cache, then the caller will also fallback.
-    fn select_victim(&self) -> Result<usize>;
-}
