+++ kernel/src/vm/vmo/mod.rs
@@ -26,8 +26,6 @@ mod pager;
 
 pub use options::VmoOptions;
 pub use pager::Pager;
-#[cfg(not(baseline_asterinas))]
-pub use pager::PagerOQueues;
 
 /// Virtual Memory Objects (VMOs) are a type of capability that represents a
 /// range of memory pages.
